requirements.txt
# Phase 1 — install these now
alpaca-py==0.28.0
python-dotenv==1.0.0
pytz

# Phase 2 (Kafka) — uncomment when you reach that phase
# kafka-python==2.0.2

# Phase 3 (Processing)
# pandas==2.2.0
# numpy==1.26.0

# Phase 4 (Detection)
# scikit-learn==1.4.0
# joblib==1.3.2

# Phase 5 (Storage)
# psycopg2-binary==2.9.9
# sqlalchemy==2.0.25
# pyarrow==15.0.0

# Phase 6 (API)
# fastapi==0.109.0
# uvicorn==0.27.0
# pydantic==2.6.0

# Phase 7 (Dashboard)
# streamlit==1.31.0

# Phase 8 (Alerting)
# requests==2.31.0
